Add day-3 project for Star 1
This commit adds the solution for day-3, star 1. It does a reduce over
the rows, and builds a list of hash-maps containing the symbol
encountered, and position of that symbol. It then counts the occurances
of \# (trees).

When the reduce is started, it is assumed that a lack of previous
'positions' means we start in the top left corner.