Refactor day-3, star-1 solution to take variable column speed
The exercise for star-2 requires a variable speed - rather than come up
with a totally unique solution, we want to base the second solution
closely off the first.

This commit adds a take-nth to the input given to a new function
(tree-count), which returns the count to be printed. New tests have been
written to assert the sled returns the expected tree count.