Add day-3 project for Star 2
This commit adds the solution for day-3, star-w<F5<F4>2. The majority of
the code is identical to star-1. The difference is -main, where instead,
tree-count is called for every input, and then multiplied together.

This commit also fixes the README.md for day1, which was missing the
'#' in its title.