Add day-1 star-2 solution for Advent of Code 2021
This commit adds the solution for star 2 on day 1. The solution is
largely identical to the solution of star 1, however a new function
(aggregate-3) has been created to group sets of 3 numbers together for
the three-measurement sliding window. Once this step has been done, the
same sonar-sweep is done.

Also removed a duplicate unit test in day 1.