Add day-1 answers for AoC
This commit adds the answers for star 1 and star 2 of the Advent of
Code 2022.

Rather than a Clojure Boot script, I have opted for Babashka as it makes
the exercises easier to run. More complex projects may still be
committed in the future if unit testing plays a larger part of the
solution.