Add solution for day-6 of Advent of code 2022
This commit adds the solutions for both stars for the Advent of Code.
This solution involves partitioning the input into chunks of "start
transmission packet" size, converting them into a set, and asking which
partition is exactly 'start transmission packet'. Besides the `def`, the
solutions are identical, but have been committed as separate scripts for
consistency with other approaches.