Add solution for star-1 on day-8 of Advent of Code 2022
This commit adds the solution for day-8 of the advent of code for the
first star. The approach involves parsing the input so that each tree
is a mini hashmap, which is then updated when the algorithm (a reduce)
checks for obstructing trees. In order to check columns, I transpose the
2D array and run the same row-checking logic.